Fixed a backlog of bug reports, and worked on python 2.3 compatibility:
- fixed templating filter function arguments [SF#678911]
- fixed multiselect in searching [SF#676874]
- fixed parsing of content-disposition filenames [SF#675116]
- added 'h' to roundup-server optarg list [SF#674070]
- fixed doc for db.history in anydbm and rdbms_common [SF#679221]
- fixed timelog example so it handles new issues [SF#678908]
- handle missing os.fork() [SF#681046]
- fixed roundup-reminder [SF#681042]
- fixed int assumptions about Number values [SF#677762]
- added warning filter for "FutureWarning: hex/oct constants > sys.maxint will
  return positive values..." (literal 0xffff0000 in portalocker.py)
- fixed ZPT code generating SyntaxWarning for assignment to None

+++ roundup/admin.py
@@ -16,7 +16,7 @@
 # BASIS, AND THERE IS NO OBLIGATION WHATSOEVER TO PROVIDE MAINTENANCE,
 # SUPPORT, UPDATES, ENHANCEMENTS, OR MODIFICATIONS.
 # 
-# $Id: admin.py,v 1.35 2002-10-03 06:56:28 richard Exp $
+# $Id: admin.py,v 1.36 2003-02-06 05:43:47 richard Exp $
 
 '''Administration commands for maintaining Roundup trackers.
 '''
@@ -497,7 +497,7 @@ Command help:
                 elif isinstance(proptype, hyperdb.Boolean):
                     props[key] = value.lower() in ('yes', 'true', 'on', '1')
                 elif isinstance(proptype, hyperdb.Number):
-                    props[key] = int(value)
+                    props[key] = float(value)
 
             # try the set
             try:
@@ -682,7 +682,7 @@ Command help:
             elif isinstance(proptype, hyperdb.Boolean):
                 props[propname] = value.lower() in ('yes', 'true', 'on', '1')
             elif isinstance(proptype, hyperdb.Number):
-                props[propname] = int(value)
+                props[propname] = float(value)
 
         # check for the key property
         propname = cl.getkey()
